SMARTLOAN/requirements.txt
Flask
mysql-connector-python
bcrypt
reportlab
python-dateutil
Werkzeug

